P4-3091 moving spreadsheets into the service level package where it belongs.
