+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/service/spreadsheet/PriceSheetAssertions.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.pecs.jpc.spreadsheet
+package uk.gov.justice.digital.hmpps.pecs.jpc.service.spreadsheet
 
 import org.assertj.core.api.Assertions.assertThat
